#23 refactor: seperated recoil and firebase utils
