+++ src/Components/utils/firebase/signInWithGoogle.js
@@ -18,7 +18,6 @@ const signInWithGoogle = async () => {
     const { user } = await signInWithPopup(auth, googleProvider);
     const { isSuccess, isNewUser } = await checkUser(user);
     if (isSuccess && isNewUser) {
-      console.log("new user");
       await addUser(user);
     }
