+++ src/hooks/useSharableLink.js
@@ -1,5 +1,5 @@
 import { useRecoilValue } from "recoil";
-import { userInfoState } from "../Components/utils/authRecoil";
+import { userInfoState } from "../Components/utils/recoil/authRecoil";
 
 const useSharableLink = () => {
   const uid = useRecoilValue(userInfoState).uid;
